+++ src/buntime/jsc.zig
@@ -52,7 +52,7 @@ pub const CustomGetterSetter = @import("./bindings/CustomGetterSetter.zig").Cust
 pub const DOMFormData = @import("./bindings/DOMFormData.zig").DOMFormData;
 pub const DOMURL = @import("./bindings/DOMURL.zig").DOMURL;
 pub const DecodedJSValue = @import("./bindings/DecodedJSValue.zig").DecodedJSValue;
-pub const DeferredError = @import("./bindings/DeferredError.zig").DeferredError;
+pub const DeferredError = @import("./api/error/DeferredError.zig").DeferredError;
 pub const GetterSetter = @import("./bindings/GetterSetter.zig").GetterSetter;
 pub const JSArray = @import("./bindings/JSArray.zig").JSArray;
 pub const JSArrayIterator = @import("./bindings/JSArrayIterator.zig").JSArrayIterator;
@@ -73,7 +73,7 @@ pub const RefString = @import("./jsc/RefString.zig");
 pub const ScriptExecutionStatus = @import("./bindings/ScriptExecutionStatus.zig").ScriptExecutionStatus;
 pub const SourceType = @import("./bindings/SourceType.zig").SourceType;
 pub const Strong = @import("./Strong.zig");
-pub const SystemError = @import("./bindings/SystemError.zig").SystemError;
+pub const SystemError = @import("./api/error/SystemError.zig").SystemError;
 pub const URL = @import("./bindings/URL.zig").URL;
 pub const URLSearchParams = @import("./bindings/URLSearchParams.zig").URLSearchParams;
 pub const VM = @import("./bindings/VM.zig").VM;
@@ -87,11 +87,11 @@ pub const MarkedArgumentBuffer = @import("./bindings/MarkedArgumentBuffer.zig").
 pub const RegularExpression = @import("./bindings/RegularExpression.zig").RegularExpression;
 
 // JavaScript-related
-pub const Errorable = @import("./bindings/Errorable.zig").Errorable;
+pub const Errorable = @import("./api/error/Errorable.zig").Errorable;
 pub const ResolvedSource = @import("./bindings/ResolvedSource.zig").ResolvedSource;
-pub const ErrorCode = @import("./bindings/ErrorCode.zig").ErrorCode;
-pub const JSErrorCode = @import("./bindings/JSErrorCode.zig").JSErrorCode;
-pub const ZigErrorType = @import("./bindings/ZigErrorType.zig").ZigErrorType;
+pub const ErrorCode = @import("./api/error/ErrorCode.zig").ErrorCode;
+pub const JSErrorCode = @import("./api/error/JSErrorCode.zig").JSErrorCode;
+pub const ZigErrorType = @import("./api/error/ZigErrorType.zig").ZigErrorType;
 pub const Debugger = @import("./Debugger.zig");
 pub const SavedSourceMap = @import("./SavedSourceMap.zig");
 pub const VirtualMachine = @import("./VirtualMachine.zig");
